requirements.txt
Django>=4.2,<5.0
google-genai>=0.5.0
exa-py>=1.0.12
faiss-cpu>=1.7.4
langchain>=0.2,<0.3
langchain-community>=0.2,<0.3
langchain-text-splitters>=0.2,<0.3
networkx>=3.1
numpy>=1.24
rank-bm25>=0.2.2
scikit-learn>=1.3
tavily-python>=0.3.3
